python/ex-fast-api-1/pyproject.toml
[dependency-groups]
dev = [
    "fastapi>=0.127.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
